hackathon-2-TODO-App/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.127.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.29",
    "uvicorn[standard]>=0.40.0",
]
